Scrollbar.py
There are two classes, each with the same parameters (except ones referring to x/y or width/height): VerticalScrollbar and HorizontalScrollbar
Terms:
Width is perpendicular to the scrolling movement of the scrollbar.
Length is parallel to the scrolling movement of the scrollbar.

Parameters:
x0, y0, x1, y1 (int/tuple etc) - used to define the coordinates (and thus the size) of the whole object, including the increment button, the decrement button and the scrollbar itself. Recommended width of 15px.
sideButtonsHeight/Width (int) - this is the length of a single increment or decrement button thus together, both buttons will take twice this integer
buttonSpacing (int) - the number of pixels between the increment/decrement buttons and the scrollbar.
numRange (int) - the maximum value to which the scale can go.
defaultNum (int) - the default starting value for the scrollbar button on the scrollbar
interval (int/float) - the size of the interval when incrementing or decrementing. Decimals are possible but not recommended without changing the output number to visibly support decimal places.
minScrollBtnHeight/Width (int) - minimum length of the scroll bar button so that it doesn't become impossibly small and unusable.
colours (tuple/list) - list/tuple of colours for the scrollbar to use in the order of border, button face, button colour when clicked, background colour of the scrollbar, text colour of the number indicator. all colours should be in hexadecimal formal of '#nnnnnn' where n is a digit 0-F
shownNumX/Y (int) - the number of pixels from the centre (0) of the scroll bar to which the shown number is adjusted. (i suggest -20)
permanentText (boolean) - if false, shown number is only visible when the scrollbar button is held down, if true always shown. if the above is None, defaults to a 0 pixel alignment
invertScaleY/X (boolean) - for True = 0 at bottom/left max at top/right, False = 0 at top/right, max at bottom/left
lock (boolean) - set to true to lock the scale (make it unusable) by default
hoverColourAlterations (int) - the amount (in base 10) that the scrollbar should be shaded when hovered over max of 255. affects all colours use positives to make it lighter, negatives to make it darker

Functions for users:
.setRange(numRange, defaultNum=None, interval=None, minHeight=None)
- Parameters same as init, from here you can customise those parameters after instance creation.
- You can also use setDefault(defaultNum), setInterval(interval) and setMinScrollBtnHeight(minHeight) to do each of the option params individually.
.setColours(clist, border=None, buttonface=None, buttonclick=None, bg=None, text=None, restoreDefaults=False)
- Use clist to alter every colour with a list or tuple of hexadecimal colour codes (make sure you get them in the right order)
- Use individual param names to only alter one or if you are unsure of the order the tuple needs to be
- Restore defaults will restore the colours used in instance creation if set to true
.shade(alt)
- Not really intended for user use, but very simply, pass in a positive or negative int to lighten or darken all colours of the scrollbar by that amount in base 10. -255<alt<255
.getValue()
- returns the numeric value of the current position of the scrollbar button.
.getRange()
- returns the max number in case you forget.
.lock(e=None, alt=-32)
- triggered by event e but an be manually triggered
- locks the scrollbar so that it cannot be interacted with
- int alt will alter the colour shading, default is -32
.unlock(e)
- triggered by event e but can be manually triggered
- unlocks scrollbar so that it can be interacted with
.resetScale(restoreColours=False)
- resets scroll button to default position, clears up shown number and can optionally reset all the colours to the defaults specified at initialisation.